requirements.txt
Django==3.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
django-jazzmin==2.4.6